pyproject.toml
[project]
name = "twitter-bot"
version = "0.1.0"
description = "AI-powered autonomous Twitter content engine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.2",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.2",
    "typer>=0.21.1",
    "rich>=14.0.0",
    "google-genai>=1.57.0",
]

[project.scripts]
twitter-bot = "twitter_bot.cli:app"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/twitter_bot"]

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
